+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -8,6 +8,8 @@ CHANGELOG
  * Register `WebLinkBundle` automatically when the WebLink component is installed
  * Make `framework.workflows` an alias of the `workflow` configuration provided by `WorkflowBundle`
  * Register `WorkflowBundle` automatically when the Workflow component is installed
+ * Make `framework.remote_event` an alias of the `remote_event` configuration provided by `RemoteEventBundle`
+ * Register `RemoteEventBundle` automatically when the RemoteEvent component is installed
  * Deprecate `JsonPathPass`, use the one from the JsonPath component instead
  * Register `ProcessBundle` automatically when the Process component is installed
  * Register `JsonPathBundle` automatically when the JsonPath component is installed