+++ src/Symfony/Component/RemoteEvent/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 8.2
 ---
 
+ * Add `RemoteEventBundle`, which provides the `remote_event` configuration and the services previously provided by `FrameworkBundle` under `framework.remote_event`
  * Add `clicked` and `unsubscribed` SMS event names
  * Allow repeating `AsRemoteEventConsumer` attribute